Q4: Design a polling server with following requirement
The polling server provides the ability to a host to connect
to the polling server, and post a poll with options
(just a string message. Ex: are you listening to me yes (1), no (0)),
should get published to the clients. The client as well, connect
with polling server with specified port.  On received the poll request,
the client can send response and keep listening for further poll.

    1. The poll server allows 1 poll to be conducted at a time,
       on port 5050 for host,  5150 for clients
    2. The host  and few clients connects to the poll server
    3. The host receives information on no.of clients connected.
    4. The host send out the poll in plain text
    5. The poll server receives the poll query and
       send to all connected clients.
    6. The client responds back (
    7. Poll server sends back individual responses in a
       consolidated string  key/value pair “Question tag,
       client1-answer, and client2-answer”
    8. The poll server waits for 1 mins to conclude the poll
    9. The host prints the response (no need to any business
       logic – just print message as such) and waits for another
       poll , till user sends CTRL+C
    10. Client connection terminates if idle for more than 2 mins